+++ crates/ruma-common/src/api/auth_scheme.rs
@@ -5,6 +5,7 @@
 
 use as_variant::as_variant;
 use http::{header, HeaderMap};
+use serde::Deserialize;
 
 /// Trait implemented by types representing an authentication scheme used by an endpoint.
 pub trait AuthScheme: Sized {
@@ -14,6 +15,12 @@ pub trait AuthScheme: Sized {
     /// The error type returned from [`add_authentication()`](Self::add_authentication).
     type AddAuthenticationError: Into<Box<dyn std::error::Error + Send + Sync + 'static>>;
 
+    /// The authentication data that can be extracted from a request.
+    type Output;
+
+    /// The error type returned from [`extract_authentication()`](Self::extract_authentication).
+    type ExtractAuthenticationError: Into<Box<dyn std::error::Error + Send + Sync + 'static>>;
+
     /// Add this authentication scheme to the given outgoing request, if necessary.
     ///
     /// Returns an error if the endpoint requires authentication but the input doesn't provide it,
@@ -22,6 +29,14 @@ pub trait AuthScheme: Sized {
         request: &mut http::Request<T>,
         input: Self::Input<'_>,
     ) -> Result<(), Self::AddAuthenticationError>;
+
+    /// Extract the data of this authentication scheme from the given incoming request.
+    ///
+    /// Returns an error if the endpoint requires authentication but the request doesn't provide it,
+    /// or if the output fails to deserialize to the proper format.
+    fn extract_authentication<T: AsRef<[u8]>>(
+        request: &http::Request<T>,
+    ) -> Result<Self::Output, Self::ExtractAuthenticationError>;
 }
 
 /// No authentication is performed.
@@ -34,6 +49,8 @@ pub struct NoAuthentication;
 impl AuthScheme for NoAuthentication {
     type Input<'a> = SendAccessToken<'a>;
     type AddAuthenticationError = header::InvalidHeaderValue;
+    type Output = ();
+    type ExtractAuthenticationError = std::convert::Infallible;
 
     fn add_authentication<T: AsRef<[u8]>>(
         request: &mut http::Request<T>,
@@ -43,6 +60,13 @@ impl AuthScheme for NoAuthentication {
             add_access_token_as_authorization_header(request.headers_mut(), access_token)?;
         }
 
+        Ok(())
+    }
+
+    /// Since this endpoint doesn't expect any authentication, this is a noop.
+    fn extract_authentication<T: AsRef<[u8]>>(
+        _request: &http::Request<T>,
+    ) -> Result<(), Self::ExtractAuthenticationError> {
         Ok(())
     }
 }
@@ -57,6 +81,9 @@ pub struct AccessToken;
 impl AuthScheme for AccessToken {
     type Input<'a> = SendAccessToken<'a>;
     type AddAuthenticationError = AddRequiredTokenError;
+    /// The access token.
+    type Output = String;
+    type ExtractAuthenticationError = ExtractTokenError;
 
     fn add_authentication<T: AsRef<[u8]>>(
         request: &mut http::Request<T>,
@@ -66,6 +93,12 @@ impl AuthScheme for AccessToken {
             .get_required_for_endpoint()
             .ok_or(AddRequiredTokenError::MissingAccessToken)?;
         Ok(add_access_token_as_authorization_header(request.headers_mut(), token)?)
+    }
+
+    fn extract_authentication<T: AsRef<[u8]>>(
+        request: &http::Request<T>,
+    ) -> Result<String, Self::ExtractAuthenticationError> {
+        extract_bearer_or_query_token(request)?.ok_or(ExtractTokenError::MissingAccessToken)
     }
 }
 
@@ -79,6 +112,9 @@ pub struct AccessTokenOptional;
 impl AuthScheme for AccessTokenOptional {
     type Input<'a> = SendAccessToken<'a>;
     type AddAuthenticationError = header::InvalidHeaderValue;
+    /// The access token, if any.
+    type Output = Option<String>;
+    type ExtractAuthenticationError = ExtractTokenError;
 
     fn add_authentication<T: AsRef<[u8]>>(
         request: &mut http::Request<T>,
@@ -89,6 +125,12 @@ impl AuthScheme for AccessTokenOptional {
         }
 
         Ok(())
+    }
+
+    fn extract_authentication<T: AsRef<[u8]>>(
+        request: &http::Request<T>,
+    ) -> Result<Option<String>, Self::ExtractAuthenticationError> {
+        extract_bearer_or_query_token(request)
     }
 }
 
@@ -103,6 +145,9 @@ pub struct AppserviceToken;
 impl AuthScheme for AppserviceToken {
     type Input<'a> = SendAccessToken<'a>;
     type AddAuthenticationError = AddRequiredTokenError;
+    /// The appservice token.
+    type Output = String;
+    type ExtractAuthenticationError = ExtractTokenError;
 
     fn add_authentication<T: AsRef<[u8]>>(
         request: &mut http::Request<T>,
@@ -112,6 +157,12 @@ impl AuthScheme for AppserviceToken {
             .get_required_for_appservice()
             .ok_or(AddRequiredTokenError::MissingAccessToken)?;
         Ok(add_access_token_as_authorization_header(request.headers_mut(), token)?)
+    }
+
+    fn extract_authentication<T: AsRef<[u8]>>(
+        request: &http::Request<T>,
+    ) -> Result<String, Self::ExtractAuthenticationError> {
+        extract_bearer_or_query_token(request)?.ok_or(ExtractTokenError::MissingAccessToken)
     }
 }
 
@@ -126,6 +177,9 @@ pub struct AppserviceTokenOptional;
 impl AuthScheme for AppserviceTokenOptional {
     type Input<'a> = SendAccessToken<'a>;
     type AddAuthenticationError = header::InvalidHeaderValue;
+    /// The appservice token, if any.
+    type Output = Option<String>;
+    type ExtractAuthenticationError = ExtractTokenError;
 
     fn add_authentication<T: AsRef<[u8]>>(
         request: &mut http::Request<T>,
@@ -137,6 +191,12 @@ impl AuthScheme for AppserviceTokenOptional {
 
         Ok(())
     }
+
+    fn extract_authentication<T: AsRef<[u8]>>(
+        request: &http::Request<T>,
+    ) -> Result<Option<String>, Self::ExtractAuthenticationError> {
+        extract_bearer_or_query_token(request)
+    }
 }
 
 /// Add the given access token as an `Authorization` HTTP header to the given map.
@@ -146,6 +206,55 @@ fn add_access_token_as_authorization_header(
 ) -> Result<(), header::InvalidHeaderValue> {
     headers.insert(header::AUTHORIZATION, format!("Bearer {token}").try_into()?);
     Ok(())
+}
+
+/// Extract the access token from the `Authorization` HTTP header or the query string of the given
+/// request.
+fn extract_bearer_or_query_token<T>(
+    request: &http::Request<T>,
+) -> Result<Option<String>, ExtractTokenError> {
+    if let Some(token) = extract_bearer_token_from_authorization_header(request.headers())? {
+        return Ok(Some(token));
+    }
+
+    if let Some(query) = request.uri().query() {
+        Ok(extract_access_token_from_query(query)?)
+    } else {
+        Ok(None)
+    }
+}
+
+/// Extract the value of the `Authorization` HTTP header with a `Bearer` scheme.
+fn extract_bearer_token_from_authorization_header(
+    headers: &HeaderMap,
+) -> Result<Option<String>, ExtractTokenError> {
+    const EXPECTED_START: &str = "bearer ";
+
+    let Some(value) = headers.get(header::AUTHORIZATION) else {
+        return Ok(None);
+    };
+
+    let value = value.to_str()?;
+
+    if value.len() < EXPECTED_START.len()
+        || !value[..EXPECTED_START.len()].eq_ignore_ascii_case(EXPECTED_START)
+    {
+        return Err(ExtractTokenError::InvalidAuthorizationScheme);
+    }
+
+    Ok(Some(value[EXPECTED_START.len()..].to_owned()))
+}
+
+/// Extract the `access_token` from the given query string.
+fn extract_access_token_from_query(
+    query: &str,
+) -> Result<Option<String>, serde_html_form::de::Error> {
+    #[derive(Deserialize)]
+    struct AccessTokenDeHelper {
+        access_token: Option<String>,
+    }
+
+    serde_html_form::from_str::<AccessTokenDeHelper>(query).map(|helper| helper.access_token)
 }
 
 /// An enum to control whether an access token should be added to outgoing requests
@@ -204,4 +313,25 @@ pub enum AddRequiredTokenError {
     /// Failed to convert the authentication to a header value.
     #[error(transparent)]
     IntoHeader(#[from] header::InvalidHeaderValue),
+}
+
+/// An error that can occur when extracting an [`AuthScheme`] that expects an access token.
+#[derive(Debug, thiserror::Error)]
+#[non_exhaustive]
+pub enum ExtractTokenError {
+    /// No access token was found, but the endpoint requires one.
+    #[error("no access token found, but this endpoint requires one")]
+    MissingAccessToken,
+
+    /// Failed to convert the header value to a UTF-8 string.
+    #[error(transparent)]
+    FromHeader(#[from] header::ToStrError),
+
+    /// The scheme of the Authorization HTTP header is invalid.
+    #[error("invalid authorization header scheme")]
+    InvalidAuthorizationScheme,
+
+    /// Failed to deserialize the query string.
+    #[error("failed to deserialize query string: {0}")]
+    FromQuery(#[from] serde_html_form::de::Error),
 }